backend/requirements.txt
django==3.2.13
psycopg2-binary
djangorestframework==3.13.1
django-environ
google-cloud-storage
django-cors-headers

djangorestframework-simplejwt
django-rest-authtoken
django-rest-auth
django-allauth